requirements.txt
coverage==7.5.2
linecache2==1.0.0
mysql-connector-python==8.4.0
six==1.16.0
traceback2==1.4.0
unittest2==1.1.0
